THE BACKEND NOW SUPPORTS TWO GET ALL TEST ROUTES
I updated the schema to use the owner property so that users can
retrieve their tests to edit, but retrieve all tests for test taking
purposes.

Their is a now a GET route that takes an email as a parameter so that a
user can retrieve only their tests for editing purposes. The existing
GET route retrieves all tests.