THE OWNER REFERENCE NOW WORKS FOR POST
I had "requiretoken" commented in the POST and PATCH routes to speed
development of my APIs.

It turns out that requiretoken is Passport functionality. It addes a
user property to the request, so that the owner property may be set
for the document about to be created. Now this all works.