I AM COMMITING A NEARLY FINAL README
I updated the README.md file following the requirements for the project.

I added a REQUIREMENTS.md file to help me track my requirements.

I added the wireframes to the project. I decided to author a single
readme so that I can share it between the front and back ends. This will
save me two hours of work.

The GitHub repo page now meets the requirements.